+++ src/sys/memory/mem.c
@@ -76,7 +76,7 @@ PetscErrorCode PetscMemoryGetCurrentUsage(PetscLogDouble *mem)
   PetscFunctionBegin;
 #if defined(PETSC_USE_PROCFS_FOR_SIZE)
 
-  PetscCall(PetscSNPrintf(proc, PETSC_STATIC_ARRAY_LENGTH(proc), "/proc/%d", (int)getpid()));
+  PetscCall(PetscSNPrintf(proc, PETSC_STATIC_ARRAY_LENGTH(proc), "/proc/%d", getpid()));
   PetscCheck((fd = open(proc, O_RDONLY)) != -1, PETSC_COMM_SELF, PETSC_ERR_FILE_OPEN, "Unable to access system file %s to get memory usage data", file);
   PetscCheck(ioctl(fd, PIOCPSINFO, &prusage) != -1, PETSC_COMM_SELF, PETSC_ERR_FILE_READ, "Unable to access system file %s to get memory usage data", file);
   *mem = (PetscLogDouble)prusage.pr_byrssize;
@@ -87,7 +87,7 @@ PetscErrorCode PetscMemoryGetCurrentUsage(PetscLogDouble *mem)
   *mem = (PetscLogDouble)(8 * fd - 4294967296); /* 2^32 - upper bits */
 
 #elif defined(PETSC_USE_PROC_FOR_SIZE) && defined(PETSC_HAVE_GETPAGESIZE)
-  PetscCall(PetscSNPrintf(proc, PETSC_STATIC_ARRAY_LENGTH(proc), "/proc/%d/statm", (int)getpid()));
+  PetscCall(PetscSNPrintf(proc, PETSC_STATIC_ARRAY_LENGTH(proc), "/proc/%d/statm", getpid()));
   PetscCheck(file = fopen(proc, "r"), PETSC_COMM_SELF, PETSC_ERR_FILE_OPEN, "Unable to access system file %s to get memory usage data", proc);
   PetscCheck(fscanf(file, "%d %d", &mm, &rss) == 2, PETSC_COMM_SELF, PETSC_ERR_SYS, "Failed to read two integers (mm and rss) from %s", proc);
   *mem = ((PetscLogDouble)rss) * ((PetscLogDouble)getpagesize());
